// src/types/produto.ts
export interface Produto {
  id: number;
  title: string;
  description: string;
  price: number;
  discountPercentage: number;
  rating: number;
  stock: number;
  brand?: string;
  category: string;
  thumbnail: string;
  images: string[];
}

export type CategoriaProduto =
  | 'beauty'
  | 'fragrances'
  | 'furniture'
  | 'groceries';